refactor: consolidate installers, cut a2a-server, drop dead scripts
Three simplifications from the codebase audit:

#2 Delete dead code + a footgun (near-zero risk)
- scripts/sync-upstream.sh: 0 references anywhere (its workflow inlines git)
- scripts/schedule_clusters.py + scripts/verify_output.py: only ever run by
  their own CI test steps, never wired into a skill or workflow; removed both
  scripts, their tests, and the two ci-tests.yml steps
- .x402books/wallets.json shipped upstream's real Base treasury/deployer
  addresses into every fork (the file's own note warned forks to delete it).
  Replaced with wallets.json.example (placeholders) + gitignored the real
  path. wallet-risk already silent-skips (WALLET_RISK_NO_WALLETS) when absent.

#4 Extract shared bash lib for the 3 installers
- New scripts/lib/skill-install.sh centralises the logic add-skill,
  install-skill-pack, and install-from-atrium each carried a copy of:
  frontmatter parse, trusted-source check, tarball fetch+extract, skills.lock
  provenance upsert, aeon.yml insert, and force-install audit logging.
- add-skill -54 lines, install-skill-pack -61 lines. The capabilities
  taxonomy stays in install-skill-pack (pack-specific + asserted by
  check-capabilities-parity.sh, which still passes).

#5 Cut apps/a2a-server, extract mcp skill-executor
- a2a-server (592 LOC, 0 tests, speculative) + add-a2a CLI + examples/a2a/*
  removed, along with A2A sections in README, SHOWCASE, CLAUDE.md,
  dependabot, aeon.yml CODE_PATHS, examples/README, and skill refs.
- New apps/mcp-server/src/skill-executor.ts holds the load/prompt/spawn/parse
  core (previously duplicated between the two servers); mcp-server imports it.
  Typechecks + builds clean.

Net: 35 files, +303 / -2126.
