--- .x402books/wallets.json
@@ -1,20 +1,0 @@
-{
-  "_note": "Fork operators: this file advertises THIS repo's agent wallets to the x402books ecosystem. Replace the addresses below with your own fork's wallets, or delete this file entirely, before deploying. Pulling upstream as-is will misattribute your agent's activity to the aaronjmars/aeon treasury.",
-  "agent": "AEON",
-  "xHandle": "@aeonframework",
-  "ecosystem": "Base",
-  "wallets": [
-    {
-      "address": "0xf1e958db7d1e4c074377946018ad645db4fb158e",
-      "role": "treasury",
-      "chain": "base",
-      "notes": "Main protocol treasury"
-    },
-    {
-      "address": "0x67976cebb5266b50a08c0dcb676e03baf305e3a2",
-      "role": "deployer",
-      "chain": "base",
-      "notes": "Contract deployer"
-    }
-  ]
-}
