+++ .x402books/wallets.json.example
@@ -1,0 +1,20 @@
+{
+  "_note": "Copy this file to .x402books/wallets.json and replace the placeholders with YOUR fork's own agent wallets before deploying. The real wallets.json is gitignored so a fork never accidentally advertises another operator's treasury to the x402books ecosystem. The wallet-risk skill reads .x402books/wallets.json and silently skips (WALLET_RISK_NO_WALLETS) when it is absent, so leaving it uncreated is safe.",
+  "agent": "YOUR_AGENT_NAME",
+  "xHandle": "@your_handle",
+  "ecosystem": "Base",
+  "wallets": [
+    {
+      "address": "0x0000000000000000000000000000000000000000",
+      "role": "treasury",
+      "chain": "base",
+      "notes": "Main protocol treasury"
+    },
+    {
+      "address": "0x0000000000000000000000000000000000000000",
+      "role": "deployer",
+      "chain": "base",
+      "notes": "Contract deployer"
+    }
+  ]
+}
