--- apps/a2a-server/.gitignore
@@ -1,2 +1,0 @@
-node_modules/
-dist/
