--- apps/a2a-server/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "aeon-a2a-server",
-  "version": "1.0.0",
-  "license": "MIT",
-  "description": "A2A Protocol Gateway — expose all Aeon skills to any AI agent framework",
-  "type": "module",
-  "main": "dist/index.js",
-  "scripts": {
-    "build": "tsc",
-    "start": "node dist/index.js",
-    "dev": "tsc && node dist/index.js"
-  },
-  "engines": {
-    "node": ">=18"
-  },
-  "devDependencies": {
-    "@types/node": "^26.0.0",
-    "typescript": "^6.0.3"
-  }
-}
