+++ skills/pr-merge/SKILL.md
@@ -101,8 +101,8 @@ Apply this rubric in order (first match wins). The rubric is conservative — wh
 
 | Bucket | Match condition | Rationale |
 |--------|-----------------|-----------|
-| **CORE_REVIEW** | Any changed path matches `install-skill-pack`, `add-skill`, `add-mcp`, `add-a2a`, `aeon`, `notify`, `notify-jsonrender`, `aeon.yml`, `generate-skills-json`, `scripts/check-capabilities-parity.sh`, `.github/workflows/aeon.yml`, `.github/workflows/chain-runner.yml`, `chain-runner.yml`, `CLAUDE.md` | The runtime executor + the things every skill depends on. A bug here ships to every fork. |
-| **INFRA_REVIEW** | Any changed path matches `.github/workflows/*.yml` (excluding `aeon.yml` already in CORE_REVIEW), `.github/actions/*`, `Dockerfile*`, `package.json` at repo root, `package-lock.json` at repo root, `apps/dashboard/package.json`, `mcp-server/package.json`, `a2a-server/package.json` | Build + CI + dependency surface. Not the executor itself but adjacent enough that the operator should look. |
+| **CORE_REVIEW** | Any changed path matches `install-skill-pack`, `scripts/lib/skill-install.sh`, `add-skill`, `add-mcp`, `aeon`, `notify`, `notify-jsonrender`, `aeon.yml`, `generate-skills-json`, `scripts/check-capabilities-parity.sh`, `.github/workflows/aeon.yml`, `.github/workflows/chain-runner.yml`, `chain-runner.yml`, `CLAUDE.md` | The runtime executor + the things every skill depends on. A bug here ships to every fork. |
+| **INFRA_REVIEW** | Any changed path matches `.github/workflows/*.yml` (excluding `aeon.yml` already in CORE_REVIEW), `.github/actions/*`, `Dockerfile*`, `package.json` at repo root, `package-lock.json` at repo root, `apps/dashboard/package.json`, `mcp-server/package.json` | Build + CI + dependency surface. Not the executor itself but adjacent enough that the operator should look. |
 | **SKILL_WARN_OR_BLOCK** | Touches any `skills/*/SKILL.md` AND `skill-scan/scan.sh` returned WARN or BLOCK on at least one of them (step 5) | A skill PR with a HIGH (BLOCK) or MEDIUM (WARN) security finding — surface explicitly. |
 | **SKILL_PASS** | Touches any `skills/*/SKILL.md` AND every scanned `SKILL.md` returned PASS | A clean skill PR. The category most likely to be safely merged once a human has read the description. |
 | **FAST_TRACK** | All changed paths match `*.md`, `*.txt`, `LICENSE*`, `docs/**`, `README*`, `SHOWCASE.md`, `ECOSYSTEM.md`, `_data/**`, `_layouts/**`, `_posts/**`, `_config.yml`, `assets/**`, `.outputs/**` | Docs/asset/data-only PR. No code path. Operator can merge on the title + a glance. |
