+++ skills/show-hn/SKILL.md
@@ -75,7 +75,7 @@ Pick a title that names exactly **one** non-obvious capability. Avoid: "framewor
 **Body** — exactly 4 paragraphs, no markdown headers inside the body, plain prose:
 
 1. **Cold open** — the lead beat from step 3 in 2–3 sentences. Concrete, dated, with a PR or commit if available. No "I'm excited to share."
-2. **What it actually does** — 4–6 sentences naming the capabilities a senior engineer would care about: schedule-driven runs on Actions, file-based memory in git, quality scoring per run, self-healing via skill-repair, MCP server + A2A gateway. Reference the README comparison table — Aeon vs Claude Code / Hermes / OpenClaw — without re-pasting it. If there's room, name one capability the senior engineer would not have guessed: the onchain security layer (`vigil` + `wallet-risk` + `vigil-revoke` — detection through revoke, Bankr-gated), or the install ecosystem (three paths: clone, `install-skill-pack`, `install-from-atrium` — the last one onchain via the Atrium marketplace), or the external-contributor inflow (skill packs landed from Nurstar / vigilcodes / HoundFlow / signa / Careful Finance / Mneme in the last 30 days). Pick ONE; do not list all three.
+2. **What it actually does** — 4–6 sentences naming the capabilities a senior engineer would care about: schedule-driven runs on Actions, file-based memory in git, quality scoring per run, self-healing via skill-repair, MCP server. Reference the README comparison table — Aeon vs Claude Code / Hermes / OpenClaw — without re-pasting it. If there's room, name one capability the senior engineer would not have guessed: the onchain security layer (`vigil` + `wallet-risk` + `vigil-revoke` — detection through revoke, Bankr-gated), or the install ecosystem (three paths: clone, `install-skill-pack`, `install-from-atrium` — the last one onchain via the Atrium marketplace), or the external-contributor inflow (skill packs landed from Nurstar / vigilcodes / HoundFlow / signa / Careful Finance / Mneme in the last 30 days). Pick ONE; do not list all three.
 3. **Honest scope** — 3–4 sentences. What it's good at (recurring background work). What it's NOT (interactive coding — keep using Claude Code for that). The "configure once, walk away" framing belongs here. Naming the boundary is what makes the rest credible.
 4. **Pointer + ask** — repo URL `https://github.com/aaronjmars/aeon`, the install one-liner (`git clone https://github.com/aaronjmars/aeon && cd aeon && ./aeon`), and a specific question for HN comments — e.g. *"What's the worst recurring-task class you've automated and abandoned because the agent kept needing you?"* Specific questions get specific replies; "feedback welcome" gets nothing.
